+++ workspace/solutions/kotlin/lib/src/test/kotlin/utils/Example.kt
@@ -20,6 +20,7 @@ package utils
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.double
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.json.int
 import kotlinx.serialization.json.jsonArray
@@ -148,8 +149,9 @@ public class Example {
             typeName: String,
         ): Any = when (typeName) {
             "kotlin.Int" -> jsonElement.jsonPrimitive.int
+            "kotlin.Double" -> jsonElement.jsonPrimitive.double
             "kotlin.IntArray" -> jsonElement.jsonArray.map { it.jsonPrimitive.int }.toIntArray()
-            else -> throw IllegalArgumentException("Unsupported type: $typeName")
+            else -> throw IllegalArgumentException("Unsupported type (json2any): $typeName")
         }
 
         /**
@@ -163,8 +165,9 @@ public class Example {
             typeName: String,
         ): JsonElement = when (typeName) {
             "kotlin.Int" -> Json.encodeToJsonElement(value as Int)
+            "kotlin.Double" -> Json.encodeToJsonElement(value as Double)
             "kotlin.IntArray" -> Json.encodeToJsonElement(value as IntArray)
-            else -> throw IllegalArgumentException("Unsupported type: $typeName")
+            else -> throw IllegalArgumentException("Unsupported type (any2json): $typeName")
         }
     }
 }