+++ workspace/solutions/rust/src/lib.rs
@@ -14,4 +14,5 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 pub mod s0;
+pub mod s1;
 pub mod utils;