+++ workspace/solutions/rust/src/s1/s8/s7/s0/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_1870_00;
+pub mod s_1870_test;
+
+pub type TSolution = fn(Vec<i32>, f64) -> i32;