+++ workspace/solutions/rust/src/s1/s8/s7/s0/s_1870_test.rs
@@ -1,0 +1,19 @@
+#[cfg(test)]
+mod tests {
+    use super::super::s_1870_00;
+    use crate::utils::{json::ValueExtension, test::t};
+
+    #[test]
+    fn test() -> () {
+        t(
+            module_path!(), //
+            &vec![s_1870_00::SOLUTION],
+            |solve, input| {
+                serde_json::Value::from(solve(
+                    input[0].to_i32_array().unwrap(), //
+                    input[1].as_f64().unwrap(),
+                ))
+            },
+        );
+    }
+}